requirements.txt
playwright==1.44.0
pytest-bdd==7.1.2
python-dotenv==1.0.1
pytest-playwright==0.5.0
